Added typedef files for custom commands Moved test scenario interfaces into typedef file
