+++ Gateway/crds-angular.api_test/.eslintignore
@@ -1,2 +1,3 @@
 node_modules
 **/*.md
+**/*.png
