+++ Gateway/crds-angular.api_test/cypress/api_tests/LoginController/GET_authenticated_oktaauth_spec.ts
@@ -1,12 +1,12 @@
-import { getToken as getOktaToken} from "shared/authorization/okta_user_auth";
-import { Test, TestConfig, unzipTests } from "shared/test_scenario_factory";
+import { getToken as getOktaToken } from "shared/authorization/okta_user_auth";
+import { unzipTests } from "shared/test_scenario_factory";
 import { Ben } from "shared/users";
 import { mpAuthenticatedBasicAuthContract, mpAuthenticatedSchemaProperties } from "./schemas/authenticatedResponse";
 
 /** Test users cannot be authenticated in Okta Prod at this time. */
 
 //Test Data
-const testConfig: TestConfig[] = [
+const testConfig: TestFactory.TestConfig[] = [
   {
     setup: [
       {
@@ -17,7 +17,7 @@ const testConfig: TestConfig[] = [
               this.data.header = { Authorization: token }
             });
         }
-      }  
+      }
     ],
     result: {
       status: 401,
@@ -36,40 +36,40 @@ const testConfig: TestConfig[] = [
 
 describe('GET /api/authenticated', () => {
   unzipTests(testConfig)
-  .forEach((t: Test) => {
-    it(t.title, () => {
-      const authenticatedRequest: Partial<Cypress.RequestOptions> = {
-        url: "/api/authenticated",
-        method: "GET",
-        failOnStatusCode: false
-      };
+    .forEach((t: TestFactory.Test) => {
+      it(t.title, () => {
+        const authenticatedRequest: Partial<Cypress.RequestOptions> = {
+          url: "/api/authenticated",
+          method: "GET",
+          failOnStatusCode: false
+        };
 
-      t.setup() //Arrange
-        .then(() => cy.request(t.buildRequest(authenticatedRequest))) //Act
-        .verifyStatus(t.result.status) //Assert
-        .itsBody(t.result.body)
-        .verifySchema(t.result.body)
-        .verifyProperties(t.result.body);
+        t.setup() //Arrange
+          .then(() => cy.request(t.buildRequest(authenticatedRequest))) //Act
+          .verifyStatus(t.result.status) //Assert
+          .itsBody(t.result.body)
+          .verifySchema(t.result.body)
+          .verifyProperties(t.result.body);
+      });
     });
-  });
 });
 
 describe('GET /api/v1.0.0/authenticated', () => {
   unzipTests(testConfig)
-  .forEach((t: Test) => {
-    it(t.title, () => {
-      const authenticatedRequest: Partial<Cypress.RequestOptions> = {
-        url: "/api/v1.0.0/authenticated",
-        method: "GET",
-        failOnStatusCode: false
-      };
+    .forEach((t: TestFactory.Test) => {
+      it(t.title, () => {
+        const authenticatedRequest: Partial<Cypress.RequestOptions> = {
+          url: "/api/v1.0.0/authenticated",
+          method: "GET",
+          failOnStatusCode: false
+        };
 
-      t.setup() //Arrange
-        .then(() => cy.request(t.buildRequest(authenticatedRequest))) //Act
-        .verifyStatus(t.result.status) //Assert
-        .itsBody(t.result.body)
-        .verifySchema(t.result.body)
-        .verifyProperties(t.result.body);
+        t.setup() //Arrange
+          .then(() => cy.request(t.buildRequest(authenticatedRequest))) //Act
+          .verifyStatus(t.result.status) //Assert
+          .itsBody(t.result.body)
+          .verifySchema(t.result.body)
+          .verifyProperties(t.result.body);
+      });
     });
-  });
 });
