+++ Gateway/crds-angular.api_test/cypress/api_tests/LoginController/POST_versioned_request-password-reset_prodspec.ts
@@ -1,6 +1,7 @@
-import { Test, TestConfig, unzipTests } from "shared/test_scenario_factory";
+import { unzipTests } from "shared/test_scenario_factory";
 import { Ben } from "shared/users";
 import { badRequestContract, badRequestProperties } from "./schemas/badRequest";
+
 /**
  * This endpoint is an alternate name for the POST /api/requestpasswordreset endpoint.
  * Tests here are lighter tests runnable in Prod - More robust tests are run in the
@@ -8,7 +9,7 @@ import { badRequestContract, badRequestProperties } from "./schemas/badRequest";
 */
 
 // Data Setup
-const testConfig: TestConfig[] = [
+const testConfig:TestFactory.TestConfig[] = [
   {
     setup: { description: "Valid Request", data: { body: { email: Ben.email } } },
     result: { status: 200 }
@@ -34,7 +35,7 @@ const testConfig: TestConfig[] = [
 // Run Tests
 describe('POST /api/v1.0.0/request-password-reset', () => {
   unzipTests(testConfig)
-    .forEach((t: Test) => {
+    .forEach((t: TestFactory.Test) => {
       it(t.title, () => {
         const requestPasswordReset: Partial<Cypress.RequestOptions> = {
           url: "/api/v1.0.0/request-password-reset",
