+++ Gateway/crds-angular.api_test/cypress/shared/test_scenario_factory.ts
@@ -1,63 +1,7 @@
-/**
- * Schema to define test scenarios and their response concisely
- * Use the unzipTests function with a TestConfig or TestConfig[] to generate a Test[]
- */
-export interface TestConfig {
-  setup: TestSetup | TestSetup[]
-  result: TestResult;
-  /**
-   * Preferred Output values are not tested, but capture the expected output for a defect scenario, or a suggestion for improvement
-   */
-  preferredResult?: TestResult;
-}
-
-interface TestSetup {
-  /** Description of the test scenario */
-  description: string;
-  /** Store any information needed in the request here. The properties "header" and "body" will be used directly in the request. */
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  data?: Record<string, any>;
-  /** Seed the database, generate/fetch test data to store in the data property, etc. here. Must return a chainable. */
-  setup?(): Cypress.Chainable<unknown>;
-}
-
-interface TestResult {
-  /** HTTP status code expected */
-  status: number;
-  /** Content of response body expected */
-  body?: ResultBody;
-}
-
-export interface ResultBody {
-  schemas?: unknown[];
-  properties?: PropertyCompare[];
-}
-
-interface PropertyCompare {
-  name: string;
-  value: string;
-  /** If false, compares with 'includes'; defaults to exact match */
-  exactMatch?: boolean;
-}
-
-/** Standardized schema for naming, setting up data, and creating request to run in a standard Test */
-export interface Test {
-  title: string;
-  /**
-   * Use this to share data between the setup and buildRequest functions.
-   * data.header and data.body will be used to build the request
-   */
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  data: Record<string, any>;
-  setup(): Cypress.Chainable<unknown>;
-  buildRequest(request?: Partial<Cypress.RequestOptions>): Partial<Cypress.RequestOptions>;
-  result: TestResult
-}
-
-function buildTest(setupConfig: TestSetup, resultConfig: TestResult): Test {
+function buildTest(setupConfig: TestFactory.TestSetup, resultConfig: TestFactory.TestResult): TestFactory.Test {
   const title = `Given ${setupConfig.description}; Expect status ${resultConfig.status}${resultConfig.body?.schemas ? ", valid schema":""}${resultConfig.body?.properties ? ", correct property values":""}`;
 
-  const test: Test = {
+  const test: TestFactory.Test = {
     title,
     data: setupConfig.data || {},
     setup: setupConfig.setup || function() {return cy.wrap({});},
@@ -72,7 +16,7 @@ function buildTest(setupConfig: TestSetup, resultConfig: TestResult): Test {
   return test;
 }
 
-function unzipConfig(config: TestConfig): Test[] {
+function unzipConfig(config: TestFactory.TestConfig): TestFactory.Test[] {
   const setups = Array.isArray(config.setup) ? config.setup : [config.setup];
   const scenarios = setups.map(s => buildTest(s, config.result));
   return scenarios;
@@ -82,7 +26,7 @@ function unzipConfig(config: TestConfig): Test[] {
  * Converts test configurations into tests
  * @param zippedTests Configurations to be converted into runnable Test[]
  */
-export function unzipTests(zippedTests: TestConfig | TestConfig[]): Test[] {
+export function unzipTests(zippedTests: TestFactory.TestConfig | TestFactory.TestConfig[]): TestFactory.Test[] {
   const configs = Array.isArray(zippedTests) ? zippedTests : [zippedTests];
   const tests = configs.map(c => unzipConfig(c)).flat();
   return tests;
